updates to ThemeDropdown
